nlp-dnn-baselines/requirements.txt
janome
mecab-python3
tensorflow-datasets
pylint
mypy
pytest